pyproject.toml
[project]
name = "address-book"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.1",
    "fastapi>=0.128.0",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.31",
    "structlog>=25.5.0",
    "uvicorn>=0.40.0",
]
